feat: autoTarget.mjs refactor - extract exportable members
