+++ fu-roll-enhancements.mjs
@@ -1,7 +1,7 @@
 import {rollEnhancements} from './module/rolls.mjs';
 import {registerKeyBindings, registerSettings} from './module/settings.mjs';
 import {initializeTemplates} from './module/templates.mjs';
-import {registerAutoTargetHooks} from './module/autoTarget.mjs';
+import {registerAutoTargetHooks} from './module/autoTarget/autoTarget.mjs';
 import {initializeActiveEffects} from './module/effects.js';
 import {getHeaderControlsItemSheetV2} from './module/documents/RollConfiguration.js';
