+++ module/effects.js
@@ -1,4 +1,4 @@
-import {MODULE} from './helpers/module-utils.mjs';
+import {MODULE} from './helpers/utils.mjs';
 
 export function initializeActiveEffects() {
     CONFIG.statusEffects.push({
