--- module/helpers/module-utils.mjs
@@ -1,1 +1,0 @@
-export const MODULE = 'fu-roll-enhancements';
